src/main/ch08/generic/Pair.java
package main.ch08.generic;

import java.lang.reflect.Constructor;
import java.lang.reflect.InvocationTargetException;

public class Pair<T> {

    private T first;
    private T second;

    public Pair() {
        first = null;
        second = null;
    }

    public Pair(T first, T second) {
        this.first = first;
        this.second = second;
    }

    public T getFirst() {
        return first;
    }

    public void setFirst(T first) {
        this.first = first;
    }

    public T getSecond() {
        return second;
    }

    public void setSecond(T second) {
        this.second = second;
    }

    public static <T> Pair<T> makePair(Class<T> cls) {
        try {

            return new Pair<>(cls.getDeclaredConstructor().newInstance(), cls.getDeclaredConstructor().newInstance());

        } catch (IllegalAccessException | NoSuchMethodException | InvocationTargetException e) {
            e.printStackTrace();
        } catch (InstantiationException e) {
            e.printStackTrace();
        }
        return null;
    }
}

